service_django/phantom_mask/requirements.txt
aiohttp==3.7.2
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.4
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto==2.49.0
boto3==1.16.8
botocore==1.19.8
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
csscompressor==0.9.5
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-ajax-selects==2.1.0
django-amazon-ses==4.0.0
django-anymail==8.5
django-appconf==1.0.4
django-cache-machine==1.1.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-compressor==2.4
django-cors-headers==3.5.0
django-currentuser==0.5.3
django-filter==2.4.0
django-htmlmin==0.11.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-light==0.1.0.post3
django-mail-queue==3.2.4
django-mptt==0.11.0
django-mptt-admin==2.0.2
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-quill-editor==0.1.22
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-richtextfield==1.6.1
django-silk==4.1.0
django-simple-captcha==0.5.13
django-simple-history==2.12.0
django-simpleui==2021.6.2
django-smart-selects==1.5.9
django-storages==1.10.1
django-tabbed-admin==1.0.4
django-tabular-permissions==2.8
django-timezone-field==4.1.1
django-translation-flags==1.0.6
django-user-agents==0.4.0
djangogrpcframework==0.2
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docutils==0.16
drf-nested-routers==0.92.1
drf-yasg==1.20.0
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
gprof2dot==2019.11.30
grpcio==1.33.2
grpcio-tools==1.33.2
gunicorn==20.1.0
hiredis==1.1.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==2.0.0
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
Menu==3.2.2
moviepy==1.0.3
msgpack==1.0.0
multidict==5.0.0
numpy==1.19.4
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
pandas==1.1.5
Pillow==8.2.0
pluggy==0.13.1
ply==3.11
proglog==0.1.9
prompt-toolkit==3.0.8
protobuf==3.13.0
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.7
pygeohash==1.2.0
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-googlegeocoder==0.4.0
python-magic==0.4.24
python-monkey-business==1.0.0
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
redis==3.5.3
reportlab==3.5.55
requests==2.24.0
rjsmin==1.1.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.3
six==1.15.0
slimit==0.8.1
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
tqdm==4.61.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.11
user-agents==2.2.0
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.2
zipp==3.4.0
zope.interface==5.1.2
